Move Bootstrap theme example to /Debug/
