+++ src/FurCoNZ.Web/Areas/Debug/Views/_ViewStart.cshtml
@@ -1,0 +1,3 @@
+﻿@{
+    Layout = "_Layout";
+}
